+++ package.json
@@ -29,6 +29,7 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/testing-library__jest-dom": "5.14.5",
     "jest": "^29.4.3",
     "jest-environment-jsdom": "^29.4.3",
