+++ src/pages/index.tsx
@@ -1,143 +1,86 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import { Inter } from '@next/font/google'
-import styles from '@/styles/Home.module.css'
-import { Button, Icon } from 'semantic-ui-react'
+import React, { useEffect, useState } from 'react';
+import { Container, Grid, Header, Card, Divider } from 'semantic-ui-react';
+import { useRouter } from 'next/router';
 
-const inter = Inter({ subsets: ['latin'] })
+import PokemonDetailsModal from '@/containers/PokemonDetailsModal/PokemonDetailsModal';
+import PokemonSearch from '@/containers/PokemonSearch/PokemonSearch';
+import PokemonSearchResult from '@/components/PokemonSearchResult/PokemonSearchResult';
 
-export default function Home() {
+import { DisplayListType } from '@/shared/pokemon.type';
+import RecentlyViewedList from '@/components/RecentlyViewedList/RecentlyViewedList';
+
+type ParamType = {
+  selected?: string
+}
+
+const Pokedex = () => {
+
+  const router = useRouter();
+
+  const { selected } = router.query as ParamType;
+
+  const [displayList, setDisplayList] = useState<DisplayListType>({});
+  const [recentViewed, setRecentViewed] = useState<string[]>([])
+
+  useEffect(() => {
+    const localList = localStorage.getItem('recentViewList') || '[]';
+    setRecentViewed(JSON.parse(localList))
+  }, [])
+
+  const updateRecentView = (name: string) => {
+    //add this name on top the list and pick first 8 items
+    const updatedList = [...recentViewed];
+    const existingIndex = recentViewed.indexOf(name);
+    if (existingIndex > -1) {
+      updatedList.splice(existingIndex, 1)
+    }
+    updatedList.unshift(name);
+    const viewList = updatedList.splice(0, 8);
+    setRecentViewed(viewList);
+    localStorage.setItem('recentViewList', JSON.stringify(viewList))
+  }
+
+  const updateParams = (selected?: string) => {
+    const query: ParamType = {};
+    if (selected) {
+      query.selected = selected;
+      updateRecentView(selected)
+    }
+    router.push({
+      pathname: '/', query
+    }, undefined, { shallow: true })
+  }
+
+  const handleSearchResult = (result: DisplayListType, triggerDetails: boolean) => {
+    setDisplayList(result);
+    updateParams(triggerDetails ? result?.list && result?.list[0]?.title : '')
+  }
+
+  const handleSelect = (name?: string) => {
+    updateParams(name);
+  }
+
   return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className={styles.main}>
-        <div className={styles.description}>
-          <p>
-            Get started by editing&nbsp;
-            <code className={styles.code}>src/pages/index.tsx</code>
-          </p>
-          <div>
-            <a
-              href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              By{' '}
-              <Image
-                src="/vercel.svg"
-                alt="Vercel Logo"
-                className={styles.vercelLogo}
-                width={100}
-                height={24}
-                priority
-              />
-            </a>
-            <Button>Click Here</Button>
-          </div>
-        </div>
-        <div>
-          <Button animated>
-            <Button.Content visible>Next</Button.Content>
-            <Button.Content hidden>
-              <Icon name='arrow right' />
-            </Button.Content>
-          </Button>
-          <Button animated='vertical'>
-            <Button.Content hidden>Shop</Button.Content>
-            <Button.Content visible>
-              <Icon name='shop' />
-            </Button.Content>
-          </Button>
-          <Button animated='fade'>
-            <Button.Content visible>Sign-up for a Pro account</Button.Content>
-            <Button.Content hidden>$12.99 a month</Button.Content>
-          </Button>
-        </div>
+    <Container style={{ marginTop: '2em' }}>
+      <Header as='h1' dividing>Pokédex</Header>
 
-        <div className={styles.center}>
-          <Image
-            className={styles.logo}
-            src="/next.svg"
-            alt="Next.js Logo"
-            width={180}
-            height={37}
-            priority
-          />
-          <div className={styles.thirteen}>
-            <Image
-              src="/thirteen.svg"
-              alt="13"
-              width={40}
-              height={31}
-              priority
-            />
-          </div>
-        </div>
+      <Grid centered columns={15}>
+        <Grid.Column mobile={15} tablet={15} computer={15}>
+          <PokemonSearch onSearch={handleSearchResult} />
+        </Grid.Column>
+      </Grid>
 
-        <div className={styles.grid}>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Docs <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Find in-depth information about Next.js features and&nbsp;API.
-            </p>
-          </a>
+      <PokemonSearchResult displayList={displayList} onSelect={handleSelect} />
+      
+      <Divider hidden />
+      <Divider hidden />
 
-          <a
-            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Learn <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Learn about Next.js in an interactive course with&nbsp;quizzes!
-            </p>
-          </a>
+      <RecentlyViewedList recentViewed={recentViewed} handleSelect={handleSelect} />
 
-          <a
-            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Templates <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Discover and deploy boilerplate example Next.js&nbsp;projects.
-            </p>
-          </a>
+      <PokemonDetailsModal name={selected} onClose={updateParams} />
 
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Deploy <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Instantly deploy your Next.js site to a shareable URL
-              with&nbsp;Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-    </>
+    </Container>
   )
 }
+
+export default Pokedex
